integrated services with transaction | integrated notification

+++ finance-management-service/src/main/java/com/main/service/DebtService.java
@@ -1,5 +1,6 @@
 package com.main.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -12,7 +13,7 @@ public interface DebtService {
 
     public Debt getDebtById(int loanId); 
 
-    public Debt updateDebt(int loanId, double amount);
+    public Debt updateDebt(int loanId, BigDecimal amount);
 
     public void deleteDebt(int loanId);
